Now, we can add discount or tax to the final cost

+++ src/datamodel/TaxOrDiscount.java
@@ -1,5 +1,5 @@
 package datamodel;
 
 public interface TaxOrDiscount {
-	public double calulateTaxOrDiscount(int value);
+	public double calulateTaxOrDiscount(double value);
 }
